Texpier_Electronics_Home_Automatation_Firmware
Texpier_Electronics_Home_Automatation_Firmware

It is the firmware for Texpier Electronics' basic PCB electronic cards (prototypes). These allow the control of electrical and electronic devices through the Alexa mobile application, any remote control that wishes to be recorded to the hardware (TV, DVD, Sound system, Air Conditioning, etc.), in order to control the devices connected to the hardware. . It also allows the control of voice commands through Alexa devices, for the control of electrical and electronic devices.
The IoT cores used are: the ESP32 and the Arduino MRK 1000.